Made context menu DRY with streamItem, playlistItem, and searchResultItem
