+++ Streamus Chrome Extension.csproj
@@ -77,6 +77,7 @@
     <Content Include="src\js\common\enum\desktopNotificationDuration.js" />
     <Content Include="src\js\common\lodashMixin.js" />
     <Content Include="src\js\contentScript\interceptor.js" />
+    <Content Include="src\js\foreground\model\contextMenuAction.js" />
     <Content Include="src\js\foreground\model\mediaSourceWrapper.js" />
     <Content Include="src\js\foreground\model\sourceBufferWrapper.js" />
     <Content Include="src\js\foreground\enum\keyboardKey.js" />
